zsh: make `cabdoc` print in reverse order and to stdout
In order for the commit log to make more sense, change the order to be
historical from top to bottom (oldest first). Additionally, it's nice to be able
to specify a particular file (or other destination) for the output to go when
running the command, so don't automatically put it in `cab.txt` but rather just
leave it to stdout.