zsh: Move functions to *.zsh files
Functions were a little wonky. Some were set as executable (?!), really they
just need to be in the functions directory. Give them .zsh extensions to make it
clear that they are _not_ supposed to be executable, and remove the execute bit
from their permissions. Also now that they have .zsh extensions, get rid of the
modelines, as well as the hash bangs that they never really needed.