dotfiles: reorganize
Use stow for management, and just use the current directory as all the
dotfiles. This is way simpler.

To install with stow, cd into dotfiles (anywhere on system) and run
`stow --target=$HOME .` (that trailing period is the stow package to
install, i.e. the current directory). Stow defaults to installing in the
parent directory, so if the repo is cloned to ~/.dotfiles, you can just
run `stow .` to install in $HOME.